+++ sdk/iot/pnp/inc/az_iot_pnp_client.h
@@ -89,8 +89,8 @@ AZ_NODISCARD az_result az_iot_pnp_client_init(
 AZ_NODISCARD az_result az_iot_pnp_client_get_user_name(
     az_iot_pnp_client const* client,
     char* mqtt_user_name,
-    int32_t mqtt_user_name_size,
-    int32_t* out_mqtt_user_name_length);
+    size_t mqtt_user_name_size,
+    size_t* out_mqtt_user_name_length);
 
 /**
  * @brief Gets the MQTT client id.
@@ -108,8 +108,8 @@ AZ_NODISCARD az_result az_iot_pnp_client_get_user_name(
 AZ_NODISCARD AZ_INLINE az_result az_iot_pnp_client_get_id(
     az_iot_pnp_client const* client,
     char* mqtt_client_id,
-    int32_t mqtt_client_id_size,
-    int32_t* out_mqtt_client_id_length)
+    size_t mqtt_client_id_size,
+    size_t* out_mqtt_client_id_length)
 {
   return az_iot_hub_client_id_get(
       &client->_internal.iot_hub_client,
@@ -217,9 +217,9 @@ AZ_NODISCARD az_result az_iot_pnp_client_telemetry_get_publish_topic(
     az_iot_pnp_client const* client,
     az_span component_name,
     char* mqtt_topic,
-    int32_t mqtt_topic_size,
+    size_t mqtt_topic_size,
     void* reserved,
-    int32_t* out_mqtt_topic_length);
+    size_t* out_mqtt_topic_length);
 
 /*
  *
